url_shortener_service/pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "Simple url shortener built with python and fastapi"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "sentry-sdk>=2.35.0",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
]
